+++ tests/conftest.py
@@ -1,8 +1,11 @@
 from logging import Logger
+import httpx
 import pytest
-import os
 
 from fastapi.testclient import TestClient
+
+import os
+import json
 
 from bento_etl.logger import get_logger
 
@@ -16,6 +19,7 @@ os.environ["AUTHZ_ENABLED"] = "False"
 
 from bento_etl.config import Config, get_config
 from bento_etl.main import app
+from bento_etl import authz
 
 
 @pytest.fixture
@@ -32,3 +36,36 @@ def logger(config) -> Logger:
 def test_client():
     with TestClient(app) as client:
         yield client
+
+
+@pytest.fixture(autouse=True)
+def mock_bearer_token(monkeypatch):
+    def mock_get_bearer_token(*args, **kwargs):
+        return "MockedToken"
+
+    monkeypatch.setattr(authz, "get_bearer_token", mock_get_bearer_token)
+
+
+@pytest.fixture
+def load_phenopacket_data():
+    caller_path = os.path.dirname(__file__)
+    file_path = os.path.join(caller_path, "data/synthetic_phenopackets_v2.json")
+    with open(file_path) as f:
+        file_content = json.load(f)
+    return file_content
+
+
+@pytest.fixture
+def set_mock_for_valid_post(monkeypatch):
+    async def mock_valid_post(*args, **kwargs):
+        return httpx.Response(204)
+
+    monkeypatch.setattr(httpx.AsyncClient, "post", mock_valid_post)
+
+
+@pytest.fixture
+def set_mock_for_invalid_post(monkeypatch):
+    async def mock_invalid_post(*args, **kwargs):
+        return httpx.Response(400)
+
+    monkeypatch.setattr(httpx.AsyncClient, "post", mock_invalid_post)
